+++ src/zbus_iface.rs
@@ -176,6 +176,23 @@ impl CtrlGraphics {
         Ok(())
     }
 
+    /// Be notified when the dgpu status changes:
+    /// enum GfxPower {
+    ///     Active,
+    ///     Suspended,
+    ///     Off,
+    ///     AsusDisabled,
+    ///     AsusMuxDiscreet,
+    ///     Unknown,
+    /// }
+    #[dbus_interface(signal)]
+    pub async fn notify_gfx_status(
+        signal_ctxt: &SignalContext<'_>,
+        status: &GfxPower,
+ 
+    ) -> zbus::Result<()> {
+    }
+    
     /// Recieve a notification if the graphics mode changes and to which mode
     #[dbus_interface(signal)]
     async fn notify_gfx(signal_ctxt: &SignalContext<'_>, vendor: &GfxMode) -> zbus::Result<()> {}